+++ MapboxSearch/offline/src/main/java/com/mapbox/search/offline/OfflineSearchEngineSettings.kt
@@ -134,8 +134,6 @@ public class OfflineSearchEngineSettings @JvmOverloads constructor(
     internal companion object {
 
         val DEFAULT_ENDPOINT_URI: URI = URI.create("https://api.mapbox.com")
-        const val DEFAULT_DATASET = "mbx-gen2"
-        const val DEFAULT_VERSION = ""
 
         private fun defaultTileStore() = TileStore.create()
     }